+++ server/sonar-server/src/test/java/org/sonar/server/qualitygate/QgateProjectFinderTest.java
@@ -150,7 +150,7 @@ public class QgateProjectFinderTest {
     // User can only see project 1
     dbTester.users().insertProjectPermissionOnUser(user, UserRole.USER, project1);
 
-    userSession.logIn(user.getLogin()).setUserId(user.getId().intValue());
+    userSession.logIn(user.getLogin()).setUserId(user.getId());
     Association result = underTest.find(
       builder()
         .gateId(Long.toString(qGate.getId()))
