+++ server/sonar-server/src/test/java/org/sonar/server/setting/ws/ResetActionTest.java
@@ -339,7 +339,7 @@ public class ResetActionTest {
   private void assertUserPropertyExists(String key, UserDto user) {
     assertThat(dbClient.propertiesDao().selectByQuery(PropertyQuery.builder()
       .setKey(key)
-      .setUserId(user.getId().intValue())
+      .setUserId(user.getId())
       .build(),
       dbSession)).isNotEmpty();
   }
